Update recommended CUDA precision to bfloat16 for Colab runtime

+++ tools/colab_train_helpers.py
@@ -8,7 +8,7 @@ from src.model.hybrid_ngiml import HybridNGIMLConfig
 from src.model.losses import MultiStageLossConfig
 from src.data.config import AugmentationConfig
 
-def _recommended_cuda_precision(default="fp16"):
+def _recommended_cuda_precision(default="bf16"):
     """Return recommended CUDA precision for Colab runtime. Extend as needed."""
     return default
